Consider US G0 Character Set ANSI-escape code as 0-width (like colors)
The ANSI escape code "^[(B" is used in terminal for changing the
character set, and is output by some terminal-coloring libraries as part
of color-resetting code (for example, by blessed). It's also present in
"sgr0" code of terminfo from some terminals.

The bug was that prettytable didn't interpret it as 0-width
non-printable sequence but as the sequence of individual characters,
breaking the width of the column.
Like colors codes, it should be ignored when computing the width of a
cell.

Closes #129.